fixed a bunch of bugs with simsearch interface
